requirements.txt
selenium
webdriver-manager
python-dotenv
beautifulsoup4
lxml

